git subrepo pull BuildVision2/Data/Scripts/BuildVision2/RichHudFramework
subrepo:
  subdir:   "BuildVision2/Data/Scripts/BuildVision2/RichHudFramework"
  merged:   "1a90709"
upstream:
  origin:   "git@github.com:ZachHembree/RichHudFramework.Client.git"
  branch:   "master"
  commit:   "1a90709"
git-subrepo:
  version:  "0.4.1"
  origin:   "???"
  commit:   "???"

+++ BuildVision2/Data/Scripts/BuildVision2/RichHudFramework/Shared/UI/HUD/HudElementEnums.cs
@@ -99,7 +99,7 @@ namespace RichHudFramework
         /// <summary>
         /// Used to determine text alignment.
         /// </summary>
-        public enum TextAlignment : int
+        public enum TextAlignment : byte
         {
             Left = 0,
             Center = 1,
